requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
numpy==2.0.1
    # via -r requirements.in
pillow==10.4.0
    # via -r requirements.in
pygame==2.6.0
    # via -r requirements.in
